+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Yii 2 ActiveRecord Dynamic Attribute extension Change Log
 -----------------------
 
 - Enh: Usage of deprecated `yii\base\InvalidParamException` changed to `yii\base\InvalidArgumentException` one (klimov-paul)
+- Enh #1: `JsonExpressionSerializer` created providing support for `yii\db\JsonExpression` usage (klimov-paul)
 
 
 1.0.1, November 3, 2017